Updates:
- Added the Node-Intelligent-DataRouter project and enhanced it to have two specific action consumerURLRelay.js and consumerTopicRelay.js actions.
- Enhanced environment variables references for better implementation.
